certification_materials/requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
vertexai>=1.38.0
jupyter>=1.0.0
ipykernel>=6.0.0
pandas>=1.5.0
numpy>=1.21.0 